split api calls out from index route
